+++ app.py
@@ -1,27 +1,5 @@
-from flask import Flask, jsonify, request
-from services import web_scraping as ws, telegram
-import json
+from flask import Flask
+from src.controllers.search_controller import blueprint
 
 app = Flask(__name__)
-
-@app.route("/search", methods=['GET'])
-def search_by_city_district():
-    res = []
-    city = request.args.get('city')
-    districts_list = request.args.getlist('district')
-    if(districts_list != [] and city != None):
-        for district in districts_list:
-            ad = ws.search_last_ad(city, district)
-            res.append(json.loads(ad))
-    else:
-        ad = ws.search_last_ad()
-        res.append(json.loads(ad))
-    return jsonify(res)        
-
-@app.route("/<string:city>/districts", methods=['GET'])
-def find_all_districts(city):
-    return jsonify(ws.get_districts(city))
-
-@app.route("/sendAd", methods=['POST'])
-def send_ad():
-    return telegram.send_ad(request.data)
+app.register_blueprint(blueprint, url_prefix="/search")
